layout: do not set pointer-events to none
This is almost certainly a breaking change at this point, but will be
worth it in the long run.  Specifically it will probably break
people's existing `overlay` and `dropdownPanel`-based layouts, and
anything else that positions instances directly atop other instances.
However it is our opinion that globally setting pointer-events to none
to aid in those particular cases is letting the tail wag the dog, and
we would be much better off adding special code into `overlays`,
`dropdownPanel`, and others as needed, than we would be surprising the
user time and again by having their pointer-events turned off even on
perfectly benign parts of the page.

This may change again in the future as we settle on the Right Way to
handle pointer events.